+++ src/main/java/com/rmj/guanzongroup/sidebarmenus/controller/PurchaseOrder_HistoryCarController.java
@@ -266,7 +266,7 @@ public class PurchaseOrder_HistoryCarController implements Initializable, Screen
                 tfRO.setText(CustomCommonUtil.setDecimalValueToIntegerFormat(lnRO));
                 tfBO.setText(CustomCommonUtil.setDecimalValueToIntegerFormat(lnBO));
                 tfQOH.setText(CustomCommonUtil.setDecimalValueToIntegerFormat(lnQOH));
-                tfCost.setText(CustomCommonUtil.setIntegerValueToDecimalFormat(poPurchasingController.PurchaseOrder().Detail(pnTblDetailRow).getUnitPrice(), true));
+                tfCost.setText(CustomCommonUtil.setIntegerValueToDecimalFormat(poPurchasingController.PurchaseOrder().Detail(pnTblDetailRow).getUnitPrice().doubleValue(), true));
                 tfRequestQuantity.setText(CustomCommonUtil.setDecimalValueToIntegerFormat(lnRequestQuantity));
                 tfOrderQuantity.setText(String.valueOf(poPurchasingController.PurchaseOrder().Detail(pnTblDetailRow).getQuantity().intValue()));
             }
